+++ src/core/jsproxy.c
@@ -34,7 +34,6 @@ typedef struct
 {
   PyObject_HEAD
   JsRef js;
-  bool awaited; // for promises
 } JsProxy;
 // clang-format on
 
@@ -74,9 +73,9 @@ JsProxy_GetAttr(PyObject* self, PyObject* attr)
   PyErr_Clear();
 
   bool success = false;
-  JsRef idresult;
+  JsRef idresult = 0;
   // result:
-  PyObject* pyresult;
+  PyObject* pyresult = NULL;
 
   const char* key = PyUnicode_AsUTF8(attr);
   FAIL_IF_NULL(key);
@@ -196,25 +195,23 @@ static PyObject*
 JsProxy_IterNext(PyObject* o)
 {
   JsProxy* self = (JsProxy*)o;
+  JsRef idresult = NULL;
+  PyObject* result = NULL;
 
-  JsRef idresult = hiwire_next(self->js);
-  if (idresult == NULL) {
-    return NULL;
+  int done = hiwire_next(self->js, &idresult);
+  FAIL_IF_MINUS_ONE(done);
+  // If there was no "value", "idresult" will be jsundefined
+  // so pyvalue will be set to Py_None.
+  result = js2python(idresult);
+  FAIL_IF_NULL(result);
+  if (done) {
+    PyErr_SetObject(PyExc_StopIteration, result);
+    Py_CLEAR(result);
   }
 
-  JsRef iddone = hiwire_get_member_string(idresult, "done");
-  int done = hiwire_nonzero(iddone);
-  hiwire_decref(iddone);
-
-  PyObject* pyvalue = NULL;
-  if (!done) {
-    JsRef idvalue = hiwire_get_member_string(idresult, "value");
-    pyvalue = js2python(idvalue);
-    hiwire_decref(idvalue);
-  }
-
-  hiwire_decref(idresult);
-  return pyvalue;
+finally:
+  hiwire_CLEAR(idresult);
+  return result;
 }
 
 static Py_ssize_t
@@ -319,13 +316,6 @@ JsProxy_Bool(PyObject* o)
 static PyObject*
 JsProxy_Await(JsProxy* self, PyObject* _args)
 {
-  // Guards
-  if (self->awaited) {
-    PyErr_SetString(PyExc_RuntimeError,
-                    "cannot reuse already awaited coroutine");
-    return NULL;
-  }
-
   if (!hiwire_is_promise(self->js)) {
     PyObject* str = JsProxy_Repr((PyObject*)self);
     const char* str_utf8 = PyUnicode_AsUTF8(str);
@@ -427,7 +417,6 @@ JsProxy_cinit(PyObject* obj, JsRef idobj)
 {
   JsProxy* self = (JsProxy*)obj;
   self->js = hiwire_incref(idobj);
-  self->awaited = false;
 }
 
 static PyObject*