+++ src/pyodide-py/pyodide/__init__.py
@@ -1,7 +1,7 @@
 from ._base import open_url, eval_code, find_imports, as_nested_list
 from ._core import JsException  # type: ignore
 from ._importhooks import JsFinder
-from .simple_web_loop import SimpleWebLoopPolicy
+from .webloop import WebLoopPolicy
 import asyncio
 import sys
 import platform
@@ -12,7 +12,7 @@ unregister_js_module = jsfinder.unregister_js_module
 sys.meta_path.append(jsfinder)  # type: ignore
 
 if platform.system() == "Emscripten":
-    asyncio.set_event_loop_policy(SimpleWebLoopPolicy())
+    asyncio.set_event_loop_policy(WebLoopPolicy())
 
 
 __version__ = "0.16.1"